requirements.txt
beautifulsoup4 
Brotli 
certifi 
charset-normalizer 
idna  
PySocks 
requests 
soupsieve 
urllib3
tqdm
numpy
sentence-transformers
faiss-cpu
